committino

+++ backend/s2/progetto/src/main/java/it/epicode/catalogo/Riviste.java
@@ -3,7 +3,7 @@ package it.epicode.catalogo;
 public class Riviste extends Catalogo{
         private Peridiocita periodicita;
 
-        public Riviste(String titolo, int ISBN, int anno, int pagine, Peridiocita peridiocita) {
+        public Riviste(String titolo, long ISBN, int anno, int pagine, Peridiocita peridiocita) {
                 super(titolo, ISBN, anno, pagine);
                 this.periodicita = peridiocita;
         }
